apps/api/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "An API for interacting with agents."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi-pagination>=0.12.32",
    "fastapi[standard]>=0.115.5",
    "openai>=1.54.4",
    "prisma>=0.15.0",
    "pydantic-settings>=2.6.1",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "anthropic>=0.39.0",
    "markdown>=3.7",
    "pypdf2>=3.0.1",
]
